chore: Updated hapi tests to node:test
