+++ test/versioned/hapi/ignoring.test.js
@@ -1,43 +1,45 @@
 /*
- * Copyright 2020 New Relic Corporation. All rights reserved.
+ * Copyright 2024 New Relic Corporation. All rights reserved.
  * SPDX-License-Identifier: Apache-2.0
  */
 
 'use strict'
 
-const test = require('tap').test
-const helper = require('../../lib/agent_helper')
+const test = require('node:test')
+const tspl = require('@matteo.collina/tspl')
+
 const API = require('../../../api')
+const helper = require('../../lib/agent_helper')
 const utils = require('./hapi-utils')
 
-test('ignoring a Hapi route', function (t) {
-  t.plan(6)
+test('ignoring a Hapi route', async (t) => {
+  const plan = tspl(t, { plan: 6 })
 
   const agent = helper.instrumentMockedAgent()
   const api = new API(agent)
   const server = utils.getServer()
 
-  t.teardown(function () {
+  t.after(function () {
     helper.unloadAgent(agent)
     return server.stop()
   })
 
   agent.on('transactionFinished', function (transaction) {
-    t.ok(transaction.ignore, 'transaction is ignored')
+    plan.ok(transaction.ignore, 'transaction is ignored')
 
-    t.notOk(agent.traces.trace, 'should have no transaction trace')
+    plan.equal(agent.traces.trace, undefined, 'should have no transaction trace')
 
     const metrics = agent.metrics._metrics.unscoped
     // loading k2 adds instrumentation metrics for packages it instruments
     const expectedMetrics = helper.isSecurityAgentEnabled(agent) ? 11 : 3
-    t.equal(
+    plan.equal(
       Object.keys(metrics).length,
       expectedMetrics,
       'only supportability metrics added to agent collection'
     )
 
     const errors = agent.errors.traceAggregator.errors
-    t.equal(errors.length, 0, 'no errors noticed')
+    plan.equal(errors.length, 0, 'no errors noticed')
   })
 
   server.route({
@@ -53,8 +55,10 @@ test('ignoring a Hapi route', function (t) {
     const port = server.info.port
     const uri = 'http://localhost:' + port + '/order/31337'
     helper.makeGetRequest(uri, function (_error, res, body) {
-      t.equal(res.statusCode, 400, 'got expected error')
-      t.same(body, { status: 'cartcartcart' }, 'got expected response')
+      plan.equal(res.statusCode, 400, 'got expected error')
+      plan.deepStrictEqual(body, { status: 'cartcartcart' }, 'got expected response')
     })
   })
+
+  await plan.completed
 })
